🎨 cleanup handler
